+++ src/nativelib/__init__.py
@@ -27,4 +27,4 @@ __all__ = (
     "NativeWriter",
 )
 __author__ = "0xMihalich"
-__version__ = "0.2.2.4"
+__version__ = "0.2.2.5"